In order to reduce ndpi_main.c file size: - Removed nDPI configuration code from ndpi_main.c and placed into ndpi_config.c - Moved some utils functions from ndpi_main.c to ndpi_utils.c
